TG-43: fix delete mehod
