+++ src/Components/Projects/List/index.module.css
@@ -5,9 +5,7 @@
   justify-content: center;
   align-items: center;
 }
-.projectsForm {
-  margin-bottom: 40px;
-  border: 3px solid black;
-  border-radius: 10px;
-  width: 80%;
+
+.LinkReset {
+  text-decoration: none;
 }
