--- src/Components/Projects/List/list.js
@@ -1,27 +1,0 @@
-import React from 'react';
-import ListItem from '../ListItem/listItem';
-
-const List = ({ list, deleteItem }) => {
-  return (
-    <div>
-      <table>
-        <thead>
-          <tr>
-            <th>Project</th>
-            <th>Start Date</th>
-            <th>Active</th>
-            <th>Client</th>
-            <th>Delete</th>
-          </tr>
-        </thead>
-        <tbody>
-          {list.map((item) => (
-            <ListItem key={item._id} listItem={item} deleteItem={deleteItem} />
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-};
-
-export default List;
