--- src/Components/Projects/ListItem/listItem.module.css
@@ -1,3 +1,0 @@
-.center {
-    text-align: center;
-}
